Fix #608 Aggiunta feature al deserializzatore dei messaggi json per lanciare eccezione quando si tenta di assegnare un numero intero ad una enum.
