+++ jars/core-beans/src/main/java/it/govpay/core/beans/JSONSerializable.java
@@ -1,19 +1,18 @@
 package it.govpay.core.beans;
 
-import java.io.InputStream;
 import java.util.Arrays;
 
 import org.openspcoop2.utils.serialization.IDeserializer;
 import org.openspcoop2.utils.serialization.ISerializer;
+import org.openspcoop2.utils.serialization.JsonJacksonSerializer;
 import org.openspcoop2.utils.serialization.SerializationConfig;
-import org.openspcoop2.utils.serialization.SerializationFactory;
-import org.openspcoop2.utils.serialization.SerializationFactory.SERIALIZATION_TYPE;
 
 import com.fasterxml.jackson.annotation.JsonFilter;
 import com.fasterxml.jackson.annotation.JsonIgnore;
 
 import it.govpay.core.exceptions.IOException;
 import it.govpay.core.utils.SimpleDateFormatUtils;
+import it.govpay.core.utils.serialization.JsonJacksonDeserializer;
 
 
 @JsonFilter(value="risultati")  
@@ -35,7 +34,7 @@ public abstract class JSONSerializable {
 				serializationConfig.setIncludes(Arrays.asList(fields.split(",")));
 				serializationConfig.setExcludes(null); 
 			}
-			ISerializer serializer = SerializationFactory.getSerializer(SERIALIZATION_TYPE.JSON_JACKSON, serializationConfig);
+			ISerializer serializer = new JsonJacksonSerializer(serializationConfig);
 			return serializer.getObject(this);
 		} catch(org.openspcoop2.utils.serialization.IOException e) {
 			throw new IOException("Errore nella serializzazione della risposta.", e);
@@ -43,38 +42,19 @@ public abstract class JSONSerializable {
 	}
 	
 	public static <T> T parse(String jsonString, Class<T> t) throws IOException  {
-		SerializationConfig serializationConfig = new SerializationConfig();
+		it.govpay.core.utils.serialization.SerializationConfig serializationConfig = new it.govpay.core.utils.serialization.SerializationConfig();
 		serializationConfig.setDf(SimpleDateFormatUtils.newSimpleDateFormatSoloData());
 		serializationConfig.setIgnoreNullValues(true);
+		serializationConfig.setFailOnNumbersForEnums(true);
 		return parse(jsonString, t, serializationConfig);
 	}
 	
-	public static <T> T parse(String jsonString, Class<T> t, SerializationConfig serializationConfig) throws IOException  {
+	public static <T> T parse(String jsonString, Class<T> t, it.govpay.core.utils.serialization.SerializationConfig serializationConfig) throws IOException  {
 		try {
-			IDeserializer deserializer = SerializationFactory.getDeserializer(SERIALIZATION_TYPE.JSON_JACKSON, serializationConfig);
+			IDeserializer deserializer = new JsonJacksonDeserializer(serializationConfig);
 			
 			@SuppressWarnings("unchecked")
 			T object = (T) deserializer.getObject(jsonString, t);
-			return object;
-		} catch(org.openspcoop2.utils.serialization.IOException e) {
-			throw new IOException(e.getMessage(), e);
-		}
-	}
-
-
-	public static <T> T read(InputStream jsonIS, Class<T> t) throws IOException  {
-		SerializationConfig serializationConfig = new SerializationConfig();
-		serializationConfig.setDf(SimpleDateFormatUtils.newSimpleDateFormatSoloData());
-		serializationConfig.setIgnoreNullValues(true);
-		return read(jsonIS, t, serializationConfig);
-	}
-	
-	public static <T> T read(InputStream jsonIS, Class<T> t, SerializationConfig serializationConfig) throws IOException  {
-		try {
-			IDeserializer deserializer = SerializationFactory.getDeserializer(SERIALIZATION_TYPE.JSON_JACKSON, serializationConfig);
-			
-			@SuppressWarnings("unchecked")
-			T object = (T) deserializer.readObject(jsonIS, t);
 			return object;
 		} catch(org.openspcoop2.utils.serialization.IOException e) {
 			throw new IOException(e.getMessage(), e);
